requirements.txt
wheel==0.41.2
setuptools==68.0.0
numpy==1.24.3
scikit-learn==1.0.2
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
flask==3.0.2
gunicorn==21.2.0
python-telegram-bot[job-queue]==20.8
aiohttp==3.9.3
googlemaps==4.2.0
unidecode==1.3.8
python-dotenv==1.0.1
openai==0.28.1
lingua-language-detector==1.3.2
transformers==4.31.0
sentence-transformers==2.2.2
deep-translator==1.11.4
pytz==2024.1
requests==2.31.0
pandas==1.5.3
deeppavlov==1.1.0
tqdm==4.66.2
torch==1.13.1
geopy==2.4.1
